+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- [#13](https://github.com/zendframework/ZendService_ReCaptcha/pull/13) adds support for PHP 7.1 and 7.2.
+- [zendframework/zendservice-recaptcha#13](https://github.com/zendframework/ZendService_ReCaptcha/pull/13) adds support for PHP 7.1 and 7.2.
 
 ### Changed
 
@@ -18,7 +18,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Removed
 
-- [#13](https://github.com/zendframework/ZendService_ReCaptcha/pull/13) removes support for HHVM.
+- [zendframework/zendservice-recaptcha#13](https://github.com/zendframework/ZendService_ReCaptcha/pull/13) removes support for HHVM.
 
 ### Fixed
 
@@ -28,7 +28,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- [#11](https://github.com/zendframework/ZendService_ReCaptcha/pull/11) migrates to
+- [zendframework/zendservice-recaptcha#11](https://github.com/zendframework/ZendService_ReCaptcha/pull/11) migrates to
   version 2 of the ReCaptcha API.
 
 ### Deprecated